server/ocr/requirements.txt
paddleocr>=2.7.0
paddlepaddle>=2.5.0

